+++ backend/src/main/java/com/hoppy/app/meeting/service/MeetingInquiryServiceImpl.java
@@ -1,6 +1,6 @@
 package com.hoppy.app.meeting.service;
 
-import com.hoppy.app.meeting.common.Category;
+import com.hoppy.app.meeting.Category;
 import com.hoppy.app.meeting.domain.Meeting;
 import com.hoppy.app.meeting.repository.MeetingRepository;
 import java.util.List;
